+++ wp_api/src/api_client.rs
@@ -35,6 +35,7 @@ use crate::{
             search_endpoint::{SearchRequestBuilder, SearchRequestExecutor},
             site_settings_endpoint::{SiteSettingsRequestBuilder, SiteSettingsRequestExecutor},
             taxonomies_endpoint::{TaxonomiesRequestBuilder, TaxonomiesRequestExecutor},
+            template_parts_endpoint::{TemplatePartsRequestBuilder, TemplatePartsRequestExecutor},
             template_revisions_endpoint::{
                 TemplateRevisionsRequestBuilder, TemplateRevisionsRequestExecutor,
             },
@@ -75,6 +76,7 @@ pub struct WpApiRequestBuilder {
     site_settings: Arc<SiteSettingsRequestBuilder>,
     taxonomies: Arc<TaxonomiesRequestBuilder>,
     terms: Arc<TermsRequestBuilder>,
+    template_parts: Arc<TemplatePartsRequestBuilder>,
     template_revisions: Arc<TemplateRevisionsRequestBuilder>,
     templates: Arc<TemplatesRequestBuilder>,
     themes: Arc<ThemesRequestBuilder>,
@@ -114,6 +116,7 @@ impl WpApiRequestBuilder {
             site_settings,
             taxonomies,
             terms,
+            template_parts,
             template_revisions,
             templates,
             themes,
@@ -163,6 +166,7 @@ pub struct WpApiClient {
     site_settings: Arc<SiteSettingsRequestExecutor>,
     taxonomies: Arc<TaxonomiesRequestExecutor>,
     terms: Arc<TermsRequestExecutor>,
+    template_parts: Arc<TemplatePartsRequestExecutor>,
     template_revisions: Arc<TemplateRevisionsRequestExecutor>,
     templates: Arc<TemplatesRequestExecutor>,
     themes: Arc<ThemesRequestExecutor>,
@@ -199,6 +203,7 @@ impl WpApiClient {
             site_settings,
             taxonomies,
             terms,
+            template_parts,
             template_revisions,
             templates,
             themes,
@@ -245,6 +250,7 @@ api_client_generate_endpoint_impl!(WpApi, search);
 api_client_generate_endpoint_impl!(WpApi, site_settings);
 api_client_generate_endpoint_impl!(WpApi, taxonomies);
 api_client_generate_endpoint_impl!(WpApi, terms);
+api_client_generate_endpoint_impl!(WpApi, template_parts);
 api_client_generate_endpoint_impl!(WpApi, template_revisions);
 api_client_generate_endpoint_impl!(WpApi, templates);
 api_client_generate_endpoint_impl!(WpApi, themes);