+++ wp_api/src/lib.rs
@@ -37,6 +37,7 @@ pub mod search_results;
 pub mod site_settings;
 pub mod ssl;
 pub mod taxonomies;
+pub mod template_parts;
 pub mod template_revisions;
 pub mod templates;
 pub mod terms;